move from static discovery to "dynamic dyscovery of static models"
