+++ lib/index.js
@@ -16,8 +16,8 @@ exports.create = function (Arrow) {
     capabilities: [
       Capabilities.ConnectsToADataSource,
       Capabilities.ValidatesConfiguration,
-      Capabilities.ContainsModels,
-      // Capabilities.GeneratesModels,
+      // Capabilities.ContainsModels,
+      Capabilities.GeneratesModels,
       Capabilities.CanCreate,
       Capabilities.CanRetrieve,
       Capabilities.CanUpdate,
