+++ lib/lifecycle/connect.js
@@ -1,6 +1,6 @@
-const sdkFacade = require('./../../utils/sdkFacade')
-const transformer = require('./../../utils/transformer')
+const sdkFacade = require('./../../src/sdkFacade')
+const transformer = require('./../../src/transformer')
 exports.connect = function (next) {
-  this.twilioAPI = require('../../utils/twilioAPI')(this.config, sdkFacade(this.config), transformer)
+  this.twilioAPI = require('../../src/twilioAPI')(this.config, sdkFacade(this.config), transformer)
   next()
 }
