+++ lib/schema/fetchSchema.js
@@ -1,0 +1,4 @@
+exports.fetchSchema = function (next) {
+  // Just bypass schema step
+  return next(null, {})
+}
