+++ package.json
@@ -20,7 +20,7 @@
     "pluralize": "^3.1.0"
   },
   "devDependencies": {
-    "arrow": "^*",
+    "arrow": "*",    
     "husky": "^0.13.1",
     "nyc": "^10.1.2",    
     "request": "^2.79.0",
