+++ src/models/index.js
@@ -1,0 +1,15 @@
+module.exports = {
+  account: require('./account'),
+  address: require('./address'),
+  application: require('./application'),
+  availablePhoneNumber: require('./availablePhoneNumber'),
+  call: require('./call'),
+  conference: require('./conference'),
+  incomingPhoneNumber: require('./incomingPhoneNumber'),
+  key: require('./key'),
+  message: require('./message'),
+  outgoingCallerId: require('./outgoingCallerId'),
+  queue: require('./queue'),
+  recording: require('./recording'),
+  transcription: require('./transcription')
+}
